+++ test/test_helper.rb
@@ -305,7 +305,7 @@ module ActionDispatch
     end
 
     def sign_in_as(user)
-      sign_in user
+      sign_in user, scope: :user
     end
 
     def modern_headers(extra_headers = {})